[~] Updated some methods, and added delete function

+++ Backend/src/main/java/com/unesco/core/repositories/RoomRepository.java
@@ -3,6 +3,6 @@ package com.unesco.core.repositories;
 import com.unesco.core.entities.schedule.Room;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RoomRepository extends CrudRepository <Room, Integer> {
+public interface RoomRepository extends CrudRepository <Room, Long> {
     Room findRoomByRoom(String room);
 }
